refactor: mock 데이터 파일에 작성해두었던 타입 분리
