+++ src/lib/utils/activity_mock_data.ts
@@ -1,36 +1,3 @@
-export type ActivityResponse = {
-  activities: {
-    id: number;
-    userId: number;
-    title: string;
-    description: string;
-    category: string;
-    price: number;
-    address: string;
-    bannerImageUrl: string;
-    rating: number;
-    reviewCount: number;
-    createdAt: string;
-    updatedAt: string;
-  }[]
-  totalCount: number;
-};
-
-export type ActivityInfo = {
-  id: number;
-  userId: number;
-  title: string;
-  description: string;
-  category: string;
-  price: number;
-  address: string;
-  bannerImageUrl: string;
-  rating: number;
-  reviewCount: number;
-  createdAt: string;
-  updatedAt: string;
-};
-
 export const EXAMPLE_ACTIVITY = {
   activities: [
     {
