+++ src/types/mainPage.ts
@@ -1,0 +1,32 @@
+export type ActivityResponse = {
+  activities: {
+    id: number;
+    userId: number;
+    title: string;
+    description: string;
+    category: string;
+    price: number;
+    address: string;
+    bannerImageUrl: string;
+    rating: number;
+    reviewCount: number;
+    createdAt: string;
+    updatedAt: string;
+  }[]
+  totalCount: number;
+};
+
+export type ActivityInfo = {
+  id: number;
+  userId: number;
+  title: string;
+  description: string;
+  category: string;
+  price: number;
+  address: string;
+  bannerImageUrl: string;
+  rating: number;
+  reviewCount: number;
+  createdAt: string;
+  updatedAt: string;
+};
